users-auth-api/requirements.txt
argon2_cffi==21.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
stripe==14.1.0
typing_extensions==4.15.0
urllib3==2.6.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
PyJWT==2.9.0
python-dotenv==1.0.1
pydantic[email]==2.9.2